sema: move statement typechecking to separate file
